requirements.txt
allure-python-commons~=2.13.5
selenium~=4.25.0
pytest~=8.3.2